refactor: align service directories with ADR-015 structure
- party: Add observability/ and k8s/ directories
- position-keeping: Migrate repository/ to adapters/persistence/
- financial-accounting: Move tools.go to cmd/

Updates import paths and package declarations for position-keeping
persistence layer.
